[5/n tensor engine] ProcMesh activate() / mock slicing
Pull Request resolved: https://github.com/pytorch-labs/monarch/pull/264

See test_python_actor.py for an example.

DeviceMesh is hidden inside of ProcMesh wrapper, so that you can use ProcMesh as the way to tell the tensor engine what to do.

Currently we cannot spawn sliced proc meshes so we have to fake the slicing behavior for ProcMesh to allow it to stil be used as a sliced mesh for the purpose of the tensor engine.
ghstack-source-id: 290606281
@exported-using-ghexport

Differential Revision: [D76556227](https://our.internmc.facebook.com/intern/diff/D76556227/)